Important Background Bugfix
onInstalled listener was unnecessary.
More importantly, resetting session.isPending to false needed to happen in a 'finally' clause instead of a 'then' clause in the control function. Otherwise an error anywhere in the chain had the potential to freeze the program in a pending state.
Final change is removing '...args' and replacing with a single argument. Parcel compiles ...args into 'arguments' which is disallowed in strict mode and was unnecessary regardless, as there is never more than one argument to the callback.